Save loading counter every 10 nanopubs (#29)
Optimization to reduce the number of transactions ran on the DB. We can still recover from an unexpected shutdown, it's just that we'll have to skip at most 10 nanopubs that are already loaded.

It does work, reducing the time spent on updating the loading counter from 770 ms to 120 ms, making the impact of StatusController much smaller.

I tested interrupting and restarting the loading process, it all works fine.