Fix the incremental update loop being broken
`JellyNanopubLoader.loadUpdates` was not supposed to throw any exceptions, as that would stop the scheduled executor, and in turn prevent it from scheduling the `loadUpdates` task again. I forgot that it is possible that an exception may be thrown inside the `finally {}` block there if the StatusController fails to communicate with RDF4J. I've added a try/catch here to mitigate it.

I've also plugged some more holes in logic around restoring the service after a hard restart.

Because I've noticed some inconsistencies in load counters that caused repeated loads of a single nanopub, I've added some logging to track situations like this, it will be easier to debug it later if it happens again.