Fix the incremental loader getting stuck on last nanopub
Closes #40

I can't reproduce the original issue, but it's likely an edge case where the counter is not getting saved to the DB for some reason. To mitigate that, I changed the code to also save the counter if the loader thinks it's up to date anyway, just to be sure.